pyproject.toml
[tool.poetry]
name = "prcut"
version = "0.1.0"
description = ""
authors = ["Anonymous (for now)"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
torch = "^2.4.0"
torchvision = "^0.19.0"
scipy = "^1.14.0"
scikit-learn = "^1.5.1"
torch-cluster = "^1.6.3+pt24cu121"
tqdm = "^4.66.5"
munkres = "^1.1.4"
jupyterlab-vim = "^4.1.3"
jupyter = "^1.0.0"
notebook = "^7.2.1"
tensorboard = "^2.17.1"
solo-learn = {git = "https://github.com/vturrisi/solo-learn.git"}
clip = {git = "https://github.com/openai/CLIP.git"}
h5py = "^3.12.1"
gdown = "^5.2.0"

[[tool.poetry.source]]
name = "torch_cluster"
url = "https://data.pyg.org/whl/torch-2.4.0+cu121.html"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
